requirements.txt
streamlit
pymongo[srv]
Pillow
numpy
opencv-python-headless